Added support for URLs as resource paths for ImageItem.
Instead of just allowing local images to be rendered inside the report
you can put an URL (Check QNetworkAccessManager::supportedSchemes())
that points directly to an image as the resource path and the image will
be downloaded from the server (whether local or on the internet).

This allows for conserving disk space and using the functionality that
the file system provides as opposed to saving the image as a byte array
inside a database (which is not recommended at all).